SEIP-Training-To-Do-App-ToDoShow-Menu-Added

+++ app/src/main/java/com/example/mytodoapp/ToDoFragment.kt
@@ -2,6 +2,7 @@ package com.example.mytodoapp
 
 import android.os.Bundle
 import android.view.*
+import android.widget.RadioButton
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.asLiveData
@@ -10,6 +11,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.mytodoapp.adapter.TodoAdapter
 import com.example.mytodoapp.databinding.FragmentToDoBinding
+import com.example.mytodoapp.dialogs.AlertDialog
 import com.example.mytodoapp.entities.Todo
 import com.example.mytodoapp.models.TodoModel
 import com.example.mytodoapp.prefdata.LoginPreference
@@ -19,7 +21,8 @@ class ToDoFragment : Fragment() {
     private val modelView : TodoModel by activityViewModels()
     private lateinit var binding: FragmentToDoBinding
     private lateinit var loginPreference: LoginPreference
-
+    private var userId : Long = 0
+    val adapter = TodoAdapter(::onComplete)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setHasOptionsMenu(true)
@@ -34,6 +37,60 @@ class ToDoFragment : Fragment() {
             R.id.itemLogout -> {
                 lifecycle.coroutineScope.launch {
                     loginPreference.setLoginStatus(false, context = requireContext())
+                }
+            }
+            R.id.itemCompleted -> {
+                lifecycle.coroutineScope.launch {
+                    modelView
+                        .getAllTodoByCompleteCheck(userId, true)
+                        .observe(viewLifecycleOwner) {
+                            adapter.submitList(it)
+                        }
+                }
+            }
+            R.id.itemIncompleted -> {
+                lifecycle.coroutineScope.launch {
+                    modelView
+                        .getAllTodoByCompleteCheck(userId, false)
+                        .observe(viewLifecycleOwner) {
+                            adapter.submitList(it)
+                        }
+                }
+            }
+            R.id.itemLow -> {
+                lifecycle.coroutineScope.launch {
+                    modelView
+                        .getAllTodoByPriorityCheck(userId, "Low")
+                        .observe(viewLifecycleOwner) {
+                            adapter.submitList(it)
+                        }
+                }
+            }
+            R.id.itemNormal -> {
+                lifecycle.coroutineScope.launch {
+                    modelView
+                        .getAllTodoByPriorityCheck(userId, "Normal")
+                        .observe(viewLifecycleOwner) {
+                            adapter.submitList(it)
+                        }
+                }
+            }
+            R.id.itemHigh -> {
+                lifecycle.coroutineScope.launch {
+                    modelView
+                        .getAllTodoByPriorityCheck(userId, "High")
+                        .observe(viewLifecycleOwner) {
+                            adapter.submitList(it)
+                        }
+                }
+            }
+            R.id.itemAll -> {
+                lifecycle.coroutineScope.launch {
+                    modelView
+                        .getAllTodoByUser(userId)
+                        .observe(viewLifecycleOwner) {
+                            adapter.submitList(it)
+                        }
                 }
             }
         }
@@ -51,20 +108,23 @@ class ToDoFragment : Fragment() {
                 findNavController().navigate(R.id.action_toDoFragment_to_loginFragment)
             }
         }
-        val adapter = TodoAdapter(::onComplete)
+
 
         binding.todoRecycleView.layoutManager = LinearLayoutManager(activity)
         binding.todoRecycleView.adapter = adapter
 
+
         loginPreference.userIdFlow.asLiveData().observe(viewLifecycleOwner){
+            userId = it
             modelView.getAllTodoByUser(it).observe(viewLifecycleOwner){
                 val todoList = mutableListOf<Todo>()
-                for(todo in it){
-                    if(!todo.isDone){
-                        todoList.add(todo)
+                it.forEach {
+                    if(!it.isDone){
+                        todoList.add(it)
                     }
                 }
                 adapter.submitList(todoList)
+
             }
         }
 
@@ -83,7 +143,14 @@ class ToDoFragment : Fragment() {
                 modelView.updateTodo(todo)
             }
             "Delete" -> {
-                modelView.deleteTodo(todo)
+                AlertDialog(
+                    title = "Delete ${todo.name}?",
+                    body = "Are you sure to delete?",
+                    positiveText = "Yes",
+                    negativeText = "Cancel"
+                ){
+                    modelView.deleteTodo(todo)
+                }.show(childFragmentManager, null)
             }
         }
     }
